RDM-3063 : Move editor routing to shared case-edit dir
